+++ Engine/Engine/Code/Components/UVScrollComponent.cpp
@@ -38,7 +38,7 @@ void UVScrollComponent::Update(const float& delta)
 	const std::vector<Mesh*>& meshes = _modelComponent->GetMeshes(); 
 
 	// set the offset on all meshes
-	unsigned int count = meshes.size();
+	size_t count = meshes.size();
 	for (int i = 0; i < count; i++)
 		meshes[i]->SetUvOffset(_uvOffset);
 }
